+++ client/src/modelView/ProblemView.ts
@@ -194,7 +194,9 @@ export abstract class ProblemView<TRendererOptions, TRenderData> extends Disposa
             return viewPanel.getError()!.message;
         }
         else {
-            return getWebViewHtml(createPddlExtensionContext(this.context), this.options.content, this.options.webviewHtmlPath, viewPanel.getPanel().webview);
+            return getWebViewHtml(createPddlExtensionContext(this.context), {
+                relativePath: this.options.content, htmlFileName: this.options.webviewHtmlPath
+            }, viewPanel.getPanel().webview);
         }
     }